refactor: add repo and reporesponse type
